client/requirements.txt
websockets==12.0
aiohttp==3.9.1
asyncio==3.4.3
psutil==5.9.6
python-dotenv==1.0.0
pydantic==2.5.0
cryptography==41.0.7